+++ admin/core/Driver.php
@@ -456,7 +456,7 @@ abstract class Driver
 	/**
 	 * Returns inherited tables.
 	 *
-	 * @return list<string>
+	 * @return list<array{table: string, ns: string}>
 	 */
 	public function getInheritedTables(string $table): array
 	{
@@ -466,7 +466,7 @@ abstract class Driver
 	/**
 	 * Returns tables this table inherits from.
 	 *
-	 * @return list<string>
+	 * @return list<array{table: string, ns: string}>
 	 */
 	public function getParentTables(string $table): array
 	{
@@ -531,10 +531,11 @@ abstract class Driver
 		// MariaDB contains CHECK_CONSTRAINTS.TABLE_NAME, MySQL and PostgreSQL not.
 		return get_key_vals("SELECT c.CONSTRAINT_NAME, CHECK_CLAUSE
 FROM INFORMATION_SCHEMA.CHECK_CONSTRAINTS c
-JOIN INFORMATION_SCHEMA.TABLE_CONSTRAINTS t ON c.CONSTRAINT_SCHEMA = t.CONSTRAINT_SCHEMA AND c.CONSTRAINT_NAME = t.CONSTRAINT_NAME
+JOIN INFORMATION_SCHEMA.TABLE_CONSTRAINTS t ON c.CONSTRAINT_SCHEMA = t.CONSTRAINT_SCHEMA AND c.CONSTRAINT_NAME = t.CONSTRAINT_NAME" . ($this->connection->isMariaDB() ? " AND c.TABLE_NAME = t.TABLE_NAME" : "") . "
 WHERE c.CONSTRAINT_SCHEMA = " . q($_GET["ns"] != "" ? $_GET["ns"] : DB) . "
 AND t.TABLE_NAME = " . q($table) . "
-AND CHECK_CLAUSE NOT LIKE '% IS NOT NULL'", $this->connection); // ignore default IS NOT NULL checks in PostgreSQL
+AND t.TABLE_NAME = " . q($table) . (DIALECT == "pgsql" ? "
+AND CHECK_CLAUSE NOT LIKE '% IS NOT NULL'" : ""), $this->connection); // ignore default IS NOT NULL checks in PostgreSQL
 	}
 
 	/**