+++ admin/core/Origin.php
@@ -597,7 +597,7 @@ abstract class Origin extends Plugin
 
 	public abstract function printTablePartitions(array $partitionInfo): void;
 
-	public abstract function printInheritedTables(array $inheritedTables): void;
+	public abstract function printRelatedTables(array $tables): void;
 
 	public abstract function printTableIndexes(array $indexes, array $tableStatus): void;
 
@@ -706,6 +706,26 @@ abstract class Origin extends Plugin
 		return false;
 	}
 
+	/**
+	 * Returns server variables.
+	 *
+	 * @return list<string[]> [[$name, $value]]
+	 */
+	function getServerVariables(): array
+	{
+		return show_variables();
+	}
+
+	/**
+	 * Returns status variables.
+	 *
+	 * @return list<string[]> [[$name, $value]]
+	 */
+	function getStatusVariables(): array
+	{
+		return show_status();
+	}
+
 	public abstract function getDumpOutputs(): array;
 
 	public abstract function getDumpFormats(): array;
@@ -738,11 +758,11 @@ abstract class Origin extends Plugin
 		if ($missing != "auth") {
 			echo "<span class='version'>";
 			echo h(preg_replace('~\\.0(-|$)~', '$1', VERSION));
-			echo "<a id='version' class='version-badge' href='https://www.adminneo.org/download' " . target_blank() . " title='" . h($last_version) . "'>";
 			if ($this->config->isVersionVerificationEnabled() && $last_version && version_compare(VERSION, $last_version) < 0) {
+				echo "<a id='version' class='version-badge' href='https://www.adminneo.org/download' " . target_blank() . " title='" . h($last_version) . "'>";
 				echo icon_solo("asterisk");
+				echo "</a>";
 			}
-			echo "</a>";
 			echo "</span>\n";
 
 			if ($this->config->isVersionVerificationEnabled() && !$last_version) {