+++ admin/include/functions.inc.php
@@ -290,7 +290,7 @@ function get_rows($query, ?Connection $connection = null, $error = "<p class='er
 function unique_array(array $row, array $indexes): ?array
 {
 	foreach ($indexes as $index) {
-		if (!preg_match("~PRIMARY|UNIQUE~", $index["type"])) {
+		if (!preg_match("~PRIMARY|UNIQUE~", $index["type"]) && !$index["partial"]) {
 			continue;
 		}
 
@@ -832,12 +832,13 @@ function dump_headers(string $identifier, bool $multi_table = false): string
 * @param string[]
 */
 function dump_csv($row): void {
+	$tsv = $_POST["format"] == "tsv";
 	foreach ($row as $key => $val) {
-		if (preg_match('~["\n,;\t]|^0.|\.\d*0$~', $val) || $val === "") {
+		if (preg_match('~["\n]|^0[^.]|\.\d*0$|' . ($tsv ? '\t' : '[,;]|^$') . '~', $val)) {
 			$row[$key] = '"' . str_replace('"', '""', $val) . '"';
 		}
 	}
-	echo implode(($_POST["format"] == "csv" ? "," : ($_POST["format"] == "tsv" ? "\t" : ";")), $row) . "\r\n";
+	echo implode(($_POST["format"] == "csv" ? "," : ($tsv ? "\t" : ";")), $row) . "\r\n";
 }
 
 /** Apply SQL function
@@ -970,7 +971,7 @@ function get_random_string(): string
 }
 
 /** Format value to use in select
-* @param string|string[]
+* @param string|string[]|list<string[]>
 * @param string
 * @param ?array
 * @param ?int
@@ -979,11 +980,26 @@ function get_random_string(): string
 function select_value($val, $link, $field, $text_length) {
 	if (is_array($val)) {
 		$return = "";
-		foreach ($val as $k => $v) {
-			$return .= "<tr>"
-				. ($val != array_values($val) ? "<th>" . h($k) : "")
-				. "<td>" . select_value($v, $link, $field, $text_length)
-			;
+		if (array_filter($val, 'is_array') == array_values($val)) { // list of arrays
+			$keys = [];
+			foreach ($val as $v) {
+				$keys += array_fill_keys(array_keys($v), null);
+			}
+			foreach (array_keys($keys) as $k) {
+				$return .= "<th>" . h($k);
+			}
+			foreach ($val as $v) {
+				$return .= "<tr>";
+				foreach (array_merge($keys, $v) as $v2) {
+					$return .= "<td>" . select_value($v2, $link, $field, $text_length);
+				}
+			}
+		} else {
+			foreach ($val as $k => $v) {
+				$return .= "<tr>"
+					. ($val != array_values($val) ? "<th>" . h($k) : "")
+					. "<td>" . select_value($v, $link, $field, $text_length);
+			}
 		}
 		return "<table>$return</table>";
 	}
@@ -1004,7 +1020,7 @@ function select_value($val, $link, $field, $text_length) {
 	if (!$link) {
 		$link = Admin::get()->getFieldValueLink($val, $field);
 	}
-	$return = $field ? Admin::get()->formatFieldValue($val, $field) : $val;
+	$return = $field ? Admin::get()->formatFieldValue(Connection::get()->formatValue($val, $field), $field) : $val;
 	if ($return !== null) {
 		if (!is_utf8($return)) {
 			$return = "\0"; // htmlspecialchars of binary data returns an empty string
@@ -1056,26 +1072,37 @@ function is_mail($value): bool
 	return is_string($value) && filter_var($value, FILTER_VALIDATE_EMAIL);
 }
 
-/** Check whether the string is web URL address
-* @param string
-* @return bool
-*/
-function is_web_url($value) {
-	if (!is_string($value) || !preg_match('~^https?://~i', $value)) {
+/**
+ * Check whether the value is web URL address.
+ */
+function is_web_url($value): bool
+{
+	if (!is_string($value) || !preg_match('~^(https?:)?//~i', $value)) {
 		return false;
 	}
 
 	$components = parse_url($value);
-    if (!$components) {
-        return false;
-    }
+	if (!$components) {
+		return false;
+	}
 
-    // Encode URL path. If path was encoded already, it will be encoded twice, but we are OK with that.
-	$encodedParts = array_map('urlencode', explode('/', $components['path']));
-	$url = str_replace($components['path'], implode('/', $encodedParts), $value);
+	$url = $value;
 
-	parse_str($components['query'], $params);
-	$url = str_replace($components['query'], http_build_query($params), $url);
+	// Encode URL path. If path was encoded already, it will be encoded twice, but we are OK with that.
+	if (isset($components['path'])) {
+		$encodedParts = array_map('urlencode', explode('/', $components['path']));
+		$url = str_replace($components['path'], implode('/', $encodedParts), $url);
+	}
+
+	if (isset($components['query'])) {
+		parse_str($components['query'], $params);
+		$url = str_replace($components['query'], http_build_query($params), $url);
+	}
+
+	// FILTER_VALIDATE_URL requires a scheme, add a dummy one for //domain.tld values.
+	if (!isset($components['scheme'])) {
+		$url = "https:$url";
+	}
 
 	return (bool)filter_var($url, FILTER_VALIDATE_URL);
 }
@@ -1085,7 +1112,7 @@ function is_web_url($value) {
  */
 function is_shortable(?array $field): bool
 {
-	return $field ? preg_match('~char|text|json|lob|geometry|point|linestring|polygon|string|bytea|hstor~', $field["type"]) : false;
+	return $field && !preg_match('~' . number_type() . '|date|time|year~', $field["type"]);
 }
 
 /** Split server into host and (port or socket)