+++ admin/translations/_template.inc.php
@@ -29,6 +29,7 @@ return [
 	'Logout' => 'Xx',
 	'Logged as: %s' => 'Xx: %s',
 	'Logout successful.' => 'Xx.',
+	'hostname[:port] or :socket' => 'xx',
 	'Invalid server or credentials.' => 'Xx.',
 	'There is a space in the input password which might be the cause.' => 'Xx.',
 	'AdminNeo does not support accessing a database without a password, <a href="https://www.adminneo.org/password"%s>more information</a>.' => 'Xx, <a href="https://www.adminneo.org/password"%s>xx</a>.',